+++ src/commands/updater.ts
@@ -1,6 +1,6 @@
 import { compare, parse } from "@std/semver";
 import { FetchGitHub } from "../utils/fetch.ts";
-import { RELEASE_URL, VERSION } from "../constants.ts";
+import { RELEASE_URL, VERSIONING } from "../constants.ts";
 import type { GITHUB_RELEASE } from "../types/misc.ts";
 import { GetDateNow } from "../functions/date.ts";
 import type { TheUpdaterConstructedParams } from "./constructors/command.ts";
@@ -11,7 +11,7 @@ import type { CF_FKNODE_SCHEDULE } from "../types/config_files.ts";
 
 async function TellAboutUpdate(newVer: string): Promise<void> {
     await LogStuff(
-        `There's a new version! ${newVer}. Consider downloading it from GitHub. You're on ${VERSION}, btw.`,
+        `There's a new version! ${newVer}. Consider downloading it from GitHub. You're on ${VERSIONING.APP}, btw.`,
         "bulb",
     );
 }
@@ -26,7 +26,7 @@ export default async function TheUpdater(params: TheUpdaterConstructedParams): P
     const scheduleFilePath = await GetAppPath("SCHEDULE");
     const scheduleFileContents = parseYaml(await Deno.readTextFile(scheduleFilePath)) as CF_FKNODE_SCHEDULE;
 
-    const IsUpToDate: (tag: string) => boolean = (tag: string) => compare(parse(VERSION), parse(tag)) >= 0;
+    const IsUpToDate: (tag: string) => boolean = (tag: string) => compare(parse(VERSIONING.APP), parse(tag)) >= 0;
 
     async function CheckUpdates(): Promise<CF_FKNODE_SCHEDULE | "rl"> {
         try {
@@ -68,7 +68,7 @@ export default async function TheUpdater(params: TheUpdaterConstructedParams): P
 
     if (IsUpToDate(latestVer)) {
         if (params.silent) return;
-        await LogStuff(`You're up to date! ${ColorString(VERSION, "bright-green")} is the latest.`, "tick");
+        await LogStuff(`You're up to date! ${ColorString(VERSIONING.APP, "bright-green")} is the latest.`, "tick");
         return;
     } else {
         await TellAboutUpdate(latestVer);