+++ CHANGELOG.md
@@ -78,6 +78,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Fixed the CLI sometimes not finding your projects because of string manipulation issues.
 - Fixed Git-related commands sometimes not working because of output handling.
   - As as side effect, you now don't get to see Git's output live.
+- Fixed the CLI not being able to handle projects that were missing the `name` or `version` field in a `package.json`/`deno.json` file.
 
 ## [2.2.1] 16-01-2025